+++ test/plugin_test.dart
@@ -406,5 +406,6 @@ class MockCLI implements CLISetup {
   static const name = 'mock-cli';
 
   @override
-  Future<String> download(HostPlatform platform) => Future.value(name);
+  Future<String> download(HostPlatform platform, String directory) =>
+      Future.value(name);
 }